+++ AmazonChimeSDK/AmazonChimeSDK/session/CreateMeetingResponse.swift
@@ -19,6 +19,7 @@ import Foundation
 @objcMembers public class Meeting: NSObject {
     let externalMeetingId: String?
     let mediaPlacement: MediaPlacement
+    let meetingFeatures: MeetingFeatures
     let mediaRegion: String
     let meetingId: String
     let primaryMeetingId: String?
@@ -30,6 +31,7 @@ import Foundation
         self.init(
             externalMeetingId: externalMeetingId,
             mediaPlacement: mediaPlacement,
+            meetingFeatures: MeetingFeatures(),
             mediaRegion: mediaRegion,
             meetingId: meetingId,
             primaryMeetingId: nil)
@@ -37,11 +39,13 @@ import Foundation
 
     public init(externalMeetingId: String?,
                  mediaPlacement: MediaPlacement,
+                 meetingFeatures: MeetingFeatures,
                  mediaRegion: String,
                  meetingId: String,
                  primaryMeetingId: String?) {
         self.externalMeetingId = externalMeetingId
         self.mediaPlacement = mediaPlacement
+        self.meetingFeatures = meetingFeatures
         self.mediaRegion = mediaRegion
         self.meetingId = meetingId
         self.primaryMeetingId = primaryMeetingId
@@ -74,4 +78,40 @@ import Foundation
         self.turnControlUrl = turnControlUrl
         self.eventIngestionUrl = eventIngestionUrl
     }
+}
+
+@objcMembers public class MeetingFeatures: NSObject {
+    public let videoMaxResolution: VideoResolution
+    public let contentMaxResolution: VideoResolution
+    public override convenience init() {
+        self.init(videoMaxResolution: VideoResolution.videoResolutionHD,
+                  contentMaxResolution: VideoResolution.videoResolutionFHD)
+    }
+    public convenience init(video: String?,
+                            content: String?) {
+        let videoResolution = video ?? "hd"
+        let contentResolution = content ?? "fhd"
+        self.init(videoMaxResolution: parseMaxResolution(resolution: videoResolution),
+                  contentMaxResolution: parseMaxResolution(resolution: contentResolution))
+    }
+    public init(videoMaxResolution: VideoResolution,
+                contentMaxResolution: VideoResolution) {
+        self.videoMaxResolution = videoMaxResolution
+        self.contentMaxResolution = contentMaxResolution
+    }
+}
+
+private func parseMaxResolution(resolution: String) -> VideoResolution {
+    let maxResolution: VideoResolution
+    let lowerCaseResolution = resolution.lowercased()
+    if (lowerCaseResolution == "none") {
+        maxResolution = VideoResolution.videoDisabled
+    } else if (lowerCaseResolution == "hd") {
+        maxResolution = VideoResolution.videoResolutionHD
+    } else if (lowerCaseResolution == "fhd") {
+        maxResolution = VideoResolution.videoResolutionFHD
+    } else {
+        maxResolution = VideoResolution.videoResolutionUHD
+    }
+    return maxResolution
 }